Refactor test
separate assertion helper definition from test files, because they are not contextual lines.
